feat: comprehensive test documentation and coverage analysis
🧪 TEST SUITE OVERHAUL:
- Add comprehensive TODO.md with test coverage analysis (17% error code coverage)
- Document all 9 existing tests with detailed descriptions and expected behaviors
- Create 18 stubbed test files organized in 4 priority phases
- Identify critical gap: claim_tokens_v0 has ZERO test coverage

📋 DOCUMENTATION IMPROVEMENTS:
- Add detailed test headers explaining what each test should validate
- Mark problematic tests with specific ignore reasons for review
- Document behavior changes (e.g., auto-advancement in TestFixture)
- Provide clear implementation roadmap with priorities

🎯 CRITICAL FINDINGS:
- Phase 1: 7 claim_tokens_v0 tests needed (HIGHEST PRIORITY)
- Phase 2: 4 lifecycle/activation tests
- Phase 3: 4 admin operation tests
- Phase 4: 3 advanced testing scenarios

✅ CURRENT STATE:
- 5 tests working and properly documented
- 4 tests ignored with clear reasons
- 18 test stubs ready for implementation
- Complete error code mapping (19/23 untested)

This establishes a clear testing roadmap and dramatically improves
the visibility into what needs to be tested vs what's already covered.
