+++ crates/prism-protocol-testing/build.rs
@@ -1,35 +1,38 @@
 use std::{env, path::PathBuf, process::Command};
 
 fn main() {
-    let manifest_dir = env::var("CARGO_MANIFEST_DIR")
-        .expect("CARGO_MANIFEST_DIR not set");
-    
+    let manifest_dir = env::var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR not set");
+
     let manifest_path = PathBuf::from(&manifest_dir);
     let workspace_root = manifest_path
-        .parent()  // crates/
+        .parent() // crates/
         .unwrap()
-        .parent()  // workspace root
+        .parent() // workspace root
         .unwrap();
-    
+
     let program_manifest = workspace_root
         .join("programs")
         .join("prism-protocol")
         .join("Cargo.toml");
-    
+
     let program_src = workspace_root
         .join("programs")
         .join("prism-protocol")
         .join("src");
-    
+
     // Tell cargo to rerun this build script if the program source changes
     println!("cargo:rerun-if-changed={}", program_manifest.display());
     println!("cargo:rerun-if-changed={}", program_src.display());
-    
+
     // Build the Prism Protocol program
     let output = Command::new("cargo")
-        .args(&["build-sbf", "--manifest-path", &program_manifest.to_string_lossy()])
+        .args(&[
+            "build-sbf",
+            "--manifest-path",
+            &program_manifest.to_string_lossy(),
+        ])
         .output();
-    
+
     match output {
         Ok(output) => {
             if !output.status.success() {
@@ -45,4 +48,4 @@ fn main() {
             std::process::exit(1);
         }
     }
-} 
+}
