+++ crates/prism-protocol-testing/tests/TODO.md
@@ -1,0 +1,152 @@
+# 🧪 PRISM PROTOCOL TEST COVERAGE ANALYSIS & TODO
+
+## 📊 CURRENT STATE ✅ UPDATED
+
+### ✅ EXISTING TESTS (9 files) - **NOW ALL DOCUMENTED**
+- `test_full_campaign_flow_happy_path.rs` ✅ - End-to-end happy path (missing claims)
+- `test_mint_mismatch.rs` ✅ - Vault init with wrong mint → MintMismatch
+- `test_vault_funding_mismatch.rs` ✅ - Vault activation with wrong funding → IncorrectVaultFunding
+- `test_zero_amount_per_entitlement.rs` ✅ - Cohort with zero entitlements → InvalidEntitlements
+- `test_non_admin_cannot_activate_campaign.rs` ⚠️ #[ignore] - Wrong admin activating → CampaignAdminMismatch
+- `test_vault_initialization_before_cohort.rs` ⚠️ #[ignore] - Wrong order operations (needs review)
+- `test_vault_activation_before_initialization.rs` ⚠️ #[ignore] - Wrong order operations (needs review)
+- `test_cohort_initialization_before_campaign.rs` ⚠️ - Tests auto-advancement (behavior changed)
+- `test_campaign_fingerprint_consistency.rs` ⚠️ #[ignore] - Fingerprint validation (stub only)
+
+### 📈 ERROR CODE COVERAGE: 4/23 (17%)
+**✅ TESTED:** MintMismatch, InvalidEntitlements, IncorrectVaultFunding, CampaignAdminMismatch
+**❌ UNTESTED:** 19 error codes (see below)
+
+### 🚨 TEST STATE SUMMARY
+- **5 tests WORKING** ✅ (actually test protocol behavior)
+- **4 tests IGNORED** ⚠️ (need implementation or review)
+- **18 tests STUBBED** 📋 (Phase 1-4 below)
+- **27 total test files** (5 working + 4 ignored + 18 stubs)
+
+---
+
+## 🚨 CRITICAL GAPS
+
+### **🔥 ZERO CLAIM TESTING**
+The most critical instruction `claim_tokens_v0` has **NO TESTS WHATSOEVER**. This is the primary user-facing functionality.
+
+### **📋 INSTRUCTION COVERAGE**
+| Instruction | Tested | Priority |
+|---|---|---|
+| `claim_tokens_v0` | ❌ **0%** | **CRITICAL** |
+| `activate_campaign_v0` | ⚠️ Partial | HIGH |
+| `activate_cohort_v0` | ❌ 0% | HIGH |
+| `activate_vault_v0` | ⚠️ Partial | HIGH |
+| `reclaim_tokens_v0` | ❌ 0% | MEDIUM |
+| Campaign lifecycle | ❌ 0% | MEDIUM |
+| Error combinations | ❌ 0% | LOW |
+
+---
+
+## 🎯 TODO: PHASE 1 - CRITICAL (IMMEDIATE) ✅ STUBBED
+
+### 🚀 claim_tokens_v0 Testing Suite
+- [ ] `test_claim_tokens_happy_path.rs` ✅ - Successful claim flow
+- [ ] `test_claim_invalid_merkle_proof.rs` ✅ - Invalid proof → InvalidMerkleProof
+- [ ] `test_claim_before_go_live.rs` ✅ - Claim before slot → GoLiveDateNotReached
+- [ ] `test_claim_vault_index_out_of_bounds.rs` ✅ - Bad vault index → AssignedVaultIndexOutOfBounds
+- [ ] `test_claim_inactive_campaign.rs` ✅ - Claim from inactive → CampaignNotActive
+- [ ] `test_claim_duplicate_prevention.rs` ✅ - Prevent double claims (ClaimReceipt PDA)
+- [ ] `test_claim_numeric_overflow.rs` ✅ - Entitlements overflow → NumericOverflow
+
+---
+
+## 🎯 TODO: PHASE 2 - HIGH PRIORITY (NEXT) ✅ STUBBED
+
+### 🔄 Lifecycle & Activation Testing
+- [ ] `test_campaign_activation_edge_cases.rs` ✅ - Multiple activation scenarios
+- [ ] `test_cohort_activation_requirements.rs` ✅ - All vaults must be activated
+- [ ] `test_vault_activation_validation.rs` ✅ - Funding validation & edge cases
+- [ ] `test_campaign_status_transitions.rs` ✅ - Status change validation
+
+---
+
+## 🎯 TODO: PHASE 3 - MEDIUM PRIORITY (LATER) ✅ STUBBED
+
+### 🔧 Admin Operations & Edge Cases
+- [ ] `test_reclaim_tokens_scenarios.rs` ✅ - Token reclamation from halted campaigns
+- [ ] `test_unstoppable_campaign_behavior.rs` ✅ - Unstoppable campaign constraints
+- [ ] `test_pause_resume_workflow.rs` ✅ - Pause/resume state transitions
+- [ ] `test_count_validation.rs` ✅ - Expected vs actual count validation
+
+---
+
+## 🎯 TODO: PHASE 4 - COMPLETENESS (FINAL) ✅ STUBBED
+
+### 🧪 Advanced Testing
+- [ ] `test_edge_case_combinations.rs` ✅ - Cross-instruction interactions
+- [ ] `test_gas_optimization_verification.rs` ✅ - CU usage validation
+- [ ] ~~`test_stress_scenarios.rs`~~ ❌ - **DELETED** (Large-scale operations)
+
+---
+
+## ❌ UNTESTED ERROR CODES (19/23)
+
+### 🚨 CRITICAL PRIORITY
+- `InvalidMerkleProof` - claim_tokens_v0
+- `GoLiveDateNotReached` - claim_tokens_v0
+- `AssignedVaultIndexOutOfBounds` - claim_tokens_v0
+- `CampaignNotActive` - claim_tokens_v0
+
+### 🔥 HIGH PRIORITY
+- `NumericOverflow` - Multiple instructions
+- `CampaignAlreadyActivated` - activate_campaign_v0
+- `NotAllCohortsActivated` - activate_campaign_v0
+- `NotAllVaultsActivated` - activate_cohort_v0
+
+### ⚠️ MEDIUM PRIORITY
+- `VaultIndexOutOfBounds` - vault operations
+- `CampaignIsActive` - pause/modify operations
+- `CampaignIsUnstoppable` - pause/halt operations
+- `CampaignNotPaused` - resume_campaign_v0
+- `CampaignNotPermanentlyHalted` - reclaim_tokens_v0
+
+### 📝 LOW PRIORITY
+- `InvalidIpfsHash` - activate_campaign_v0
+- `GoLiveSlotInPast` - activate_campaign_v0
+- `NoCohortsExpected` - initialize_campaign_v0
+- `NoVaultsExpected` - initialize_cohort_v0
+- `TokenAccountOwnerMismatch` - reclaim_tokens_v0
+- `MerkleRootMismatch` - Various instructions
+- `CohortCampaignMismatch` - Various instructions
+- `CampaignFingerprintMismatch` - Various instructions
+- `InvalidStatusTransition` - Campaign lifecycle
+
+---
+
+## 💡 INSIGHTS & RECOMMENDATIONS
+
+1. **🚨 claim_tokens_v0 is the highest priority** - Zero coverage on most critical path
+2. **📊 17% error code coverage** - Most edge cases are untested
+3. **🔄 Lifecycle gaps** - Status transitions not validated
+4. **⚡ Performance blind spots** - No CU measurement/optimization
+5. **🧪 Combination testing** - No cross-instruction interactions
+6. **⚠️ Some existing tests need review** - Order dependencies unclear
+
+**Target: ~17 additional test implementations needed for comprehensive coverage**
+
+---
+
+## 📋 DEVELOPMENT NOTES
+
+- **✅ All tests now have descriptive headers** - Know what each test should do
+- **✅ Problematic tests marked with ignore reasons** - Clear what needs attention
+- All new stub tests use `#[ignore]` during development
+- One test function per file, named after the file
+- Use TestFixture for consistent setup
+- Focus on precise error code validation
+- Document expected CU consumption where relevant
+
+---
+
+## 🔄 NEXT ACTIONS
+
+1. **Implement Phase 1 claim tests** (7 files) - **HIGHEST PRIORITY**
+2. **Review ignored tests** - Fix or remove problematic tests
+3. **Implement Phase 2 lifecycle tests** (4 files)
+4. **Continue with Phase 3 & 4** as needed
