+++ crates/prism-protocol-testing/tests/test_campaign_activation_edge_cases.rs
@@ -1,0 +1,19 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test campaign activation edge cases and validation
+///
+/// Should test:
+/// - Activate campaign with all cohorts activated → success
+/// - Activate campaign with missing cohorts → NotAllCohortsActivated
+/// - Activate already activated campaign → CampaignAlreadyActivated
+/// - Activate with invalid IPFS hash (all zeros) → InvalidIpfsHash
+/// - Activate with go_live_slot in past → GoLiveSlotInPast
+/// - Verify campaign status transitions correctly to Active
+/// - Verify go_live_slot and final_db_ipfs_hash are set correctly
+#[test]
+#[ignore]
+fn test_campaign_activation_edge_cases() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement campaign activation edge cases - test all validation scenarios");
+}
