+++ crates/prism-protocol-testing/tests/test_campaign_fingerprint_consistency.rs
@@ -1,17 +1,27 @@
 use prism_protocol_testing::TestFixture;
 
-#[test] 
+/// Test campaign fingerprint consistency across operations
+///
+/// Should test:
+/// - Initialize campaign with fingerprint A
+/// - Attempt to initialize cohort with fingerprint B (mismatch)
+/// - Verify fails with CampaignFingerprintMismatch error
+/// - Attempt to initialize vault with fingerprint C (mismatch)  
+/// - Verify fails with CampaignFingerprintMismatch error
+/// - Ensure fingerprint validation is enforced across all operations
+#[test]
+#[ignore = "TODO: Not implemented - requires bypassing TestFixture to create fingerprint mismatches"]
 fn test_campaign_fingerprint_consistency() {
     let mut _test = TestFixture::default();
-    
+
     // This test would require direct instruction building to test fingerprint mismatches
     // Since TestFixture manages consistency, we'd need to bypass it for negative testing
-    // 
+    //
     // Example approach:
     // 1. Initialize campaign with fingerprint A
-    // 2. Try to initialize cohort with fingerprint B  
+    // 2. Try to initialize cohort with fingerprint B
     // 3. Should fail due to mismatch
-    
+
     println!("🚧 TODO: Implement fingerprint mismatch test");
     println!("   Requires bypassing TestFixture to create inconsistent state");
-} 
+}
