+++ crates/prism-protocol-testing/tests/test_campaign_status_transitions.rs
@@ -1,0 +1,21 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test campaign status transitions and validation
+///
+/// Should test:
+/// - Pause active campaign → success (Active → Paused)
+/// - Pause inactive campaign → CampaignNotActive
+/// - Resume paused campaign → success (Paused → Active)  
+/// - Resume active campaign → CampaignNotPaused
+/// - Halt active campaign → success (Active → PermanentlyHalted)
+/// - Halt paused campaign → success (Paused → PermanentlyHalted)
+/// - Try to pause unstoppable campaign → CampaignIsUnstoppable
+/// - Try to halt unstoppable campaign → CampaignIsUnstoppable
+/// - Invalid transitions → InvalidStatusTransition
+#[test]
+#[ignore]
+fn test_campaign_status_transitions() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement campaign status transitions - test all state changes and validation");
+}
