+++ crates/prism-protocol-testing/tests/test_claim_duplicate_prevention.rs
@@ -1,0 +1,20 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test duplicate claim prevention via ClaimReceipt PDA
+///
+/// Should test:
+/// - Set up active campaign with funded vault
+/// - Create valid claimant and merkle proof
+/// - Execute first claim_tokens_v0 successfully
+/// - Verify ClaimReceipt PDA is created
+/// - Attempt second claim_tokens_v0 with same claimant
+/// - Verify fails (ClaimReceipt PDA already exists)
+/// - Verify error is account initialization failure (PDA exists)
+/// - Verify no additional tokens transferred on second attempt
+#[test]
+#[ignore]
+fn test_claim_duplicate_prevention() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement duplicate claim prevention test - second claim should fail");
+}
