+++ crates/prism-protocol-testing/tests/test_claim_tokens_happy_path.rs
@@ -1,0 +1,19 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test the successful claim tokens flow
+///
+/// Should test:
+/// - Set up campaign → cohort → vault → activate all
+/// - Create valid merkle proof for a claimant  
+/// - Ensure go-live slot has passed
+/// - Execute claim_tokens_v0 successfully
+/// - Verify tokens transferred to claimant
+/// - Verify ClaimReceipt PDA created
+/// - Verify vault balance decreased
+#[test]
+#[ignore]
+fn test_claim_tokens_happy_path() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement successful claim tokens flow test");
+}
