+++ crates/prism-protocol-testing/tests/test_cohort_activation_requirements.rs
@@ -1,0 +1,19 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test cohort activation requirements → NotAllVaultsActivated
+///
+/// Should test:
+/// - Set up cohort with expected_vault_count = 3
+/// - Initialize all 3 vaults but only activate 2 vaults
+/// - Attempt activate_cohort_v0 → should fail with NotAllVaultsActivated
+/// - Activate the 3rd vault
+/// - Attempt activate_cohort_v0 again → should succeed
+/// - Verify cohort status and campaign cohort counters updated
+/// - Test edge case: cohort with 1 vault, activate successfully
+#[test]
+#[ignore]
+fn test_cohort_activation_requirements() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement cohort activation requirements test - should fail with NotAllVaultsActivated");
+}
