+++ crates/prism-protocol-testing/tests/test_count_validation.rs
@@ -1,0 +1,20 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test expected vs actual count validation
+///
+/// Should test:
+/// - Initialize campaign with expected_cohort_count = 0 → NoCohortsExpected  
+/// - Initialize cohort with expected_vault_count = 0 → NoVaultsExpected
+/// - Campaign activation with fewer cohorts than expected → NotAllCohortsActivated
+/// - Cohort activation with fewer vaults than expected → NotAllVaultsActivated
+/// - Verify counters increment correctly during initialization
+/// - Verify counters increment correctly during activation
+/// - Test boundary conditions (exactly matching counts)
+/// - Test overflow scenarios in counter arithmetic
+#[test]
+#[ignore]
+fn test_count_validation() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement count validation - test expected vs actual count logic");
+}
