+++ crates/prism-protocol-testing/tests/test_edge_case_combinations.rs
@@ -1,0 +1,20 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test edge case combinations and cross-instruction interactions
+///
+/// Should test:
+/// - Multiple cohorts with different merkle roots in same campaign
+/// - Multiple vaults per cohort with different funding levels  
+/// - Fingerprint consistency across all operations
+/// - Campaign with maximum number of cohorts and vaults
+/// - Mint mismatch across different instruction combinations
+/// - PDA seed collision scenarios (if possible)
+/// - Complex claim scenarios with multiple users and vaults
+/// - Stress test: campaign → multiple cohorts → multiple vaults → many claims
+#[test]
+#[ignore]
+fn test_edge_case_combinations() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement edge case combinations - test complex cross-instruction scenarios");
+}
