+++ crates/prism-protocol-testing/tests/test_gas_optimization_verification.rs
@@ -1,0 +1,20 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test compute unit (CU) usage and optimization verification
+///
+/// Should test:
+/// - Measure CU consumption for each instruction type
+/// - Verify claim_tokens_v0 CU usage is reasonable (most critical)
+/// - Compare CU usage between different instruction paths
+/// - Test CU usage with different data sizes (large merkle proofs, etc.)
+/// - Verify no regression in CU consumption after optimizations
+/// - Document baseline CU costs for performance monitoring
+/// - Test instruction combinations and cumulative CU costs
+/// - Identify most expensive operations for optimization priorities
+#[test]
+#[ignore]
+fn test_gas_optimization_verification() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement CU measurement and optimization verification");
+}
