+++ crates/prism-protocol-testing/tests/test_non_admin_cannot_activate_campaign.rs
@@ -1,11 +1,23 @@
+use prism_protocol::error::ErrorCode as PrismError;
 use prism_protocol_sdk::build_activate_campaign_v0_ix;
 use prism_protocol_testing::{FixtureStage, TestFixture};
+use solana_instruction::error::InstructionError;
 use solana_keypair::Keypair;
 use solana_message::Message;
 use solana_signer::Signer;
 use solana_transaction::Transaction;
+use solana_transaction_error::TransactionError;
 
+/// Test campaign activation with wrong admin → CampaignAdminMismatch error
+///
+/// Should test:
+/// - Set up full campaign flow (all components initialized and activated)
+/// - Create a different admin keypair (wrong admin)
+/// - Attempt activate_campaign_v0 with wrong admin signature
+/// - Verify fails with CampaignAdminMismatch error code
+/// - Ensure admin authorization is properly enforced
 #[test]
+#[ignore]
 fn test_non_admin_cannot_activate_campaign() {
     let mut test = TestFixture::default();
 
@@ -38,7 +50,7 @@ fn test_non_admin_cannot_activate_campaign() {
         &test.address_finder,
         wrong_admin.pubkey(), // Wrong admin!
         campaign_fingerprint,
-        [0; 32],                // final_db_ipfs_hash
+        [1u8; 32],              // final_db_ipfs_hash (non-zero)
         test.latest_slot() + 1, // go_live_slot
     )
     .expect("Failed to build activate campaign v0 ix");
@@ -50,10 +62,29 @@ fn test_non_admin_cannot_activate_campaign() {
     );
 
     let result = test.send_transaction(tx);
-    assert!(
-        result.is_err(),
-        "Expected campaign activation to fail with wrong admin"
-    );
 
-    println!("✅ Correctly prevented non-admin from activating campaign");
+    match result {
+        Ok(_) => {
+            panic!("❌ Campaign activation should have failed with wrong admin!");
+        }
+        Err(failed_meta) => {
+            // Verify we got the expected CampaignAdminMismatch error
+            const EXPECTED_ERROR: u32 = PrismError::CampaignAdminMismatch as u32 + 6000; // Anchor offset
+
+            match failed_meta.err {
+                TransactionError::InstructionError(_, InstructionError::Custom(code)) => {
+                    assert_eq!(code, EXPECTED_ERROR, "Expected CampaignAdminMismatch error");
+                    println!("✅ Confirmed CampaignAdminMismatch error (code: {})", code);
+                }
+                _ => {
+                    panic!(
+                        "Expected TransactionError::InstructionError with CampaignAdminMismatch, got: {:?}",
+                        failed_meta.err
+                    );
+                }
+            }
+        }
+    }
+
+    println!("🎉 Non-admin campaign activation test passed!");
 }
