+++ crates/prism-protocol-testing/tests/test_pause_resume_workflow.rs
@@ -1,0 +1,21 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test pause/resume workflow and edge cases
+///
+/// Should test:
+/// - Pause active campaign → success (Active → Paused)
+/// - Resume paused campaign → success (Paused → Active)
+/// - Pause already paused campaign → CampaignNotActive
+/// - Resume already active campaign → CampaignNotPaused
+/// - Claims fail on paused campaign → CampaignNotActive
+/// - Claims succeed after resume
+/// - Multiple pause/resume cycles
+/// - Cannot pause/resume unstoppable campaign
+/// - Verify campaign events are emitted correctly
+#[test]
+#[ignore]
+fn test_pause_resume_workflow() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement pause/resume workflow - test complete lifecycle and edge cases");
+}
