+++ crates/prism-protocol-testing/tests/test_reclaim_tokens_scenarios.rs
@@ -1,0 +1,20 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test token reclamation from halted campaigns
+///
+/// Should test:
+/// - Reclaim from permanently halted campaign → success
+/// - Reclaim from active campaign → CampaignNotPermanentlyHalted
+/// - Reclaim from paused campaign → CampaignNotPermanentlyHalted  
+/// - Reclaim from inactive campaign → CampaignNotPermanentlyHalted
+/// - Reclaim with wrong admin → CampaignAdminMismatch
+/// - Reclaim to wrong token account → TokenAccountOwnerMismatch
+/// - Verify tokens transferred to admin's account
+/// - Test partial reclamation from multiple vaults
+#[test]
+#[ignore]
+fn test_reclaim_tokens_scenarios() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement token reclamation scenarios - test all reclaim validation");
+}
