+++ crates/prism-protocol-testing/tests/test_unstoppable_campaign_behavior.rs
@@ -1,0 +1,20 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test unstoppable campaign constraints and behavior
+///
+/// Should test:
+/// - Make active campaign unstoppable → success
+/// - Make inactive campaign unstoppable → CampaignNotActive
+/// - Make already unstoppable campaign unstoppable → CampaignIsUnstoppable
+/// - Try to pause unstoppable campaign → CampaignIsUnstoppable
+/// - Try to halt unstoppable campaign → CampaignIsUnstoppable
+/// - Verify unstoppable flag is permanent (cannot be reversed)
+/// - Verify claims still work on unstoppable campaign
+/// - Test that making unstoppable is one-way operation
+#[test]
+#[ignore]
+fn test_unstoppable_campaign_behavior() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement unstoppable campaign behavior - test constraints and immutability");
+}
