+++ crates/prism-protocol-testing/tests/test_vault_activation_validation.rs
@@ -1,0 +1,19 @@
+use prism_protocol_testing::TestFixture;
+
+/// Test vault activation validation and edge cases
+///
+/// Should test:
+/// - Activate vault with correct funding → success
+/// - Activate vault with insufficient funding → IncorrectVaultFunding  
+/// - Activate vault with excess funding → IncorrectVaultFunding
+/// - Activate vault that doesn't exist → VaultIndexOutOfBounds
+/// - Activate vault twice → (should fail gracefully)
+/// - Verify vault balance validation is precise
+/// - Test with various token amounts and decimals
+#[test]
+#[ignore]
+fn test_vault_activation_validation() {
+    let mut _test = TestFixture::default();
+
+    todo!("Implement vault activation validation - test funding requirements and edge cases");
+}
