+++ crates/prism-protocol-testing/tests/test_zero_amount_per_entitlement.rs
@@ -3,6 +3,14 @@ use prism_protocol_testing::{FixtureStage, TestFixture};
 use solana_instruction::error::InstructionError;
 use solana_transaction_error::TransactionError;
 
+/// Test cohort initialization with zero amount_per_entitlement → InvalidEntitlements error
+///
+/// Should test:
+/// - Initialize campaign successfully
+/// - Set amount_per_entitlement to 0 (invalid)
+/// - Attempt cohort initialization with zero entitlements
+/// - Verify fails with InvalidEntitlements error code
+/// - Ensure zero-value validation is working properly
 #[test]
 fn test_zero_amount_per_entitlement() {
     let mut test = TestFixture::default();
@@ -24,15 +32,18 @@ fn test_zero_amount_per_entitlement() {
         Err(failed_meta) => {
             // Verify we got the expected InvalidEntitlements error
             const EXPECTED_ERROR: u32 = PrismError::InvalidEntitlements as u32 + 6000; // Anchor offset
-            
+
             match failed_meta.err {
                 TransactionError::InstructionError(_, InstructionError::Custom(code)) => {
                     assert_eq!(code, EXPECTED_ERROR, "Expected InvalidEntitlements error");
                 }
                 _ => {
-                    panic!("Expected TransactionError::InstructionError with Custom code, got: {:?}", failed_meta.err);
+                    panic!(
+                        "Expected TransactionError::InstructionError with Custom code, got: {:?}",
+                        failed_meta.err
+                    );
                 }
             }
         }
     }
-} 
+}
